Add util getAxisValues in h5web/lib
