+++ packages/lib/src/vis/utils.ts
@@ -361,3 +361,20 @@ export function getVisibleDomains(
     yVisibleDomain: [dataBottomLeft.y, dataTopRight.y],
   };
 }
+
+export function getAxisValues(
+  rawValues: NumArray | undefined,
+  axisLength: number
+): number[] {
+  if (!rawValues) {
+    return range(axisLength);
+  }
+
+  if (rawValues.length === axisLength) {
+    return toArray(rawValues);
+  }
+
+  throw new Error(
+    `Expected array to have length ${axisLength}, not ${rawValues.length}`
+  );
+}
